+++ api/calendar/subscription.ts
@@ -1,0 +1,3 @@
+import handler from "../[...path].js";
+
+export default handler;